# react/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "paths": {
...
    "/questions/question/{question_api_id}:upsert_response": {
      "post": {
        "tags": [
          "questions"
        ],
        "summary": "Upsert Response",
        "description": "Create or update a response to a question.\n\nIf a response already exists (identified by api_identifier or participant),\nupdates it. Otherwise, creates a new response.\n\nArgs:\n    question_api_id (str): API identifier of the question\n    response (ResponseUpsert): Response creation/update parameters\n    req_dep (AuthenticatedRequestDependencies): Request dependencies including database session and auth\n\nReturns:\n    Question: Updated question with the new/updated response\n\nRaises:\n    IDNotFoundException: If question or response with given API ID is not found",
        "operationId": "upsert_response_questions_question__question_api_id__upsert_response_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "question_api_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Question Api Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResponseUpsert"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionLinked"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/questions/question/{question_api_id}:upload_image": {
      "post": {
        "tags": [
          "questions"
        ],
        "summary": "Upload Image",
        "description": "Upload an image as part of a response to a question.\n\nCreates a new response if one doesn't exist for the current user,\nthen attaches the uploaded image to it.\n\nArgs:\n    question_api_id (str): API identifier of the question\n    response_image (UploadFile): Image file to upload\n    req_dep (AuthenticatedRequestDependencies): Request dependencies including database session and auth\n\nReturns:\n    Question: Updated question with the response containing the new image\n\nRaises:\n    IDNotFoundException: If question with given API ID is not found",
        "operationId": "upload_image_questions_question__question_api_id__upload_image_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "question_api_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Question Api Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_image_questions_question__question_api_id__upload_image_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionLinked"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/questions/question/{question_api_id}": {
      "delete": {
        "tags": [
          "questions"
        ],
        "summary": "Delete Question",
        "description": "Delete a question.\n\nOnly the question author or group admin can delete a question.\nQuestions with responses cannot be deleted.\n\nArgs:\n    question_api_id (str): API identifier of the question to delete\n    req_dep (AuthenticatedRequestDependencies): Request dependencies including database session and auth\n\nRaises:\n    HTTPException: If question not found, user not authorized, or question has responses",
        "operationId": "delete_question_questions_question__question_api_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "question_api_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Question Api Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/questions/question:generate": {
      "post": {
        "tags": [
          "questions"
        ],
        "summary": "Generate Question",
        "description": "Generate a question using LLM without saving it.",
        "operationId": "generate_question_questions_question_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateQuestionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateQuestionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_image_questions_question__question_api_id__upload_image_post": {
        "properties": {
          "response_image": {
            "type": "string",
            "format": "binary",
            "title": "Response Image"
          }
        },
        "type": "object",
        "required": [
          "response_image"
        ],
        "title": "Body_upload_image_questions_question__question_api_id__upload_image_post"
      },
...
      "GenerateQuestionRequest": {
        "properties": {
          "prompt": {
            "type": "string",
            "title": "Prompt"
          }
        },
        "type": "object",
        "required": [
          "prompt"
        ],
        "title": "GenerateQuestionRequest"
      },
      "GenerateQuestionResponse": {
        "properties": {
          "generated_text": {
            "type": "string",
            "title": "Generated Text"
          }
        },
        "type": "object",
        "required": [
          "generated_text"
        ],
        "title": "GenerateQuestionResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LetterStatus": {
        "type": "string",
        "enum": [
          "UPCOMING",
          "IN_PROGRESS",
          "SENT"
        ],
        "title": "LetterStatus",
        "description": "Enumeration of possible letter statuses.\nRepresents the different states a letter can be in within the system:\n- UPCOMING: Letter is scheduled but not yet active\n- IN_PROGRESS: Letter is currently active and accepting responses\n- SENT: Letter has been completed and sent"
      },
      "LetterUnlinked": {
        "properties": {
          "api_identifier": {
            "type": "string",
            "title": "Api Identifier"
          },
          "number": {
            "type": "integer",
            "title": "Number"
          },
          "status": {
            "$ref": "#/components/schemas/LetterStatus"
          },
          "send_at": {
            "type": "string",
            "format": "date-time",
            "title": "Send At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "api_identifier",
          "number",
          "status",
          "send_at",
          "created_at"
        ],
        "title": "LetterUnlinked",
        "description": "Schema for letter without linked relationships.\n\nInherits all fields from Letter but excludes relationship data."
      },
...
      "QuestionLinked": {
        "properties": {
          "question_text": {
            "type": "string",
            "title": "Question Text"
          },
          "api_identifier": {
            "type": "string",
            "title": "Api Identifier"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "letter": {
            "$ref": "#/components/schemas/LetterUnlinked"
          },
          "responses": {
            "items": {
              "$ref": "#/components/schemas/ResponseUnlinked"
            },
            "type": "array",
            "title": "Responses"
          }
        },
        "type": "object",
        "required": [
          "question_text",
          "api_identifier",
          "created_at",
          "letter",
          "responses"
        ],
        "title": "QuestionLinked",
        "description": "Question model with linked relationships.\n\nExtends the base Question model to include the parent letter and responses.\n\nAttributes:\n    letter (LetterUnlinked): Letter containing this question\n    responses (list[ResponseUnlinked]): Responses to this question"
      },
...
      "ResponseUnlinked": {
        "properties": {
          "response_text": {
            "type": "string",
            "title": "Response Text"
          },
          "api_identifier": {
            "type": "string",
            "title": "Api Identifier"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "response_text",
          "api_identifier",
          "created_at"
        ],
        "title": "ResponseUnlinked",
        "description": "Schema for response without linked relationships.\n\nInherits all fields from Response but excludes relationship data."
      },
      "ResponseUpsert": {
        "properties": {
          "response_text": {
            "type": "string",
            "title": "Response Text"
          },
          "participant_api_identifier": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Participant Api Identifier"
          },
          "api_identifier": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Identifier"
          }
        },
        "type": "object",
        "required": [
          "response_text"
        ],
        "title": "ResponseUpsert",
        "description": "Schema for creating or updating a response.\n\nAttributes:\n    response_text (str): The text content of the response\n    participant_api_identifier (Optional[str]): API identifier of the participant, optional\n    api_identifier (Optional[str]): API identifier of an existing response, optional"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "login/access-token"
          }
        }
      }
    }
  }
}